scripts/rbac/requirements.txt
azure-common
azure-core
azure-identity
azure-mgmt-authorization
azure-mgmt-core
azure-mgmt-resource
pyyaml
